:recycle: Move `connect_pipeline` from standalone function to `ResourcePool` method
