+++ CPAC/pipeline/engine/__init__.py
@@ -17,16 +17,14 @@
 """C-PAC engine."""
 
 from .engine import (
-    NodeBlock,
-    PIPELINE_BLOCKS,
     run_node_blocks,
     wrap_block,
 )
+from .nodeblock import NodeBlock
 from .resource import ResourcePool, StratPool
 
 __all__ = [
     "NodeBlock",
-    "PIPELINE_BLOCKS",
     "ResourcePool",
     "StratPool",
     "run_node_blocks",
